+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## `transcriptic` Changelog
 
 ## Unreleased
+Added
+- Row index of the Container.aliquots DataFrame object now corresponds to the well index
 
 ## v3.5.0
 Added